Clean up code for index page
Did some refactoring and code splitting for more understandable code
